Update PHP 8.1 Test and Fixtures
