+++ tests/Fixture/PHP80100/ArgumentIntersectionTypeHint.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace MockeryTest\Fixture\PHP80100;
+
+class ArgumentIntersectionTypeHint
+{
+    public function foo(IntersectionTypeHelper1Interface&IntersectionTypeHelper2Interface $foo)
+    {
+    }
+}
