+++ tests/Fixture/PHP80100/IntersectionTypeHelper1Interface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace MockeryTest\Fixture\PHP80100;
+
+interface IntersectionTypeHelper1Interface
+{
+    public function bar(): int;
+}
