+++ tests/Fixture/PHP80100/IntersectionTypeHelper2Interface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace MockeryTest\Fixture\PHP80100;
+
+interface IntersectionTypeHelper2Interface
+{
+    public function foo(): int;
+}
