+++ tests/Fixture/PHP80100/LoggerInterface.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace MockeryTest\Fixture\PHP80100;
+
+interface LoggerInterface
+{
+}
