+++ tests/Fixture/PHP80100/NullLogger.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace MockeryTest\Fixture\PHP80100;
+
+class NullLogger implements LoggerInterface
+{
+}
