+++ tests/Fixture/PHP80100/ReturnTypeWillChangeAttributeNoReturnType.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace MockeryTest\Fixture\PHP80100;
+
+class ReturnTypeWillChangeAttributeNoReturnType extends \DateTime
+{
+    #[\ReturnTypeWillChange]
+    public function getTimestamp()
+    {
+    }
+}
