+++ tests/Fixture/PHP80100/ReturnTypeWillChangeAttributeWrongReturnType.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace MockeryTest\Fixture\PHP80100;
+
+class ReturnTypeWillChangeAttributeWrongReturnType extends \DateTime
+{
+    #[\ReturnTypeWillChange]
+    public function getTimestamp() : float
+    {
+    }
+}
